Patch to remove various limits in lirc
This patch alters the stock lirc to allow GPIO RX/TX of long sequences
as found in airconditioners (f.e. Olimpia Splendid, Fujitsu, Midea, etc).

Note - you cannot use normal lircd with it as it has a limitations which
prevent long sequences of its own. You need to use the replay
utilities supplied or program the replay yourself. It is easy - just write
lists of pulse/space lengths to the actual TX device